CS 453/Homework/Homework #2/requirements.txt
dotenv==1.0.1
requests==2.31.0
google-generativeai==0.4.0